refactor: extract handleUserMessage sub-handlers from sessions.ts
